% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvrt.r
\name{convert_R2S}
\alias{convert_R2S}
\title{Convert correlation matrix to covariance matrix}
\usage{
convert_R2S(R, var)
}
\arguments{
\item{R}{Correlation matrix}

\item{var}{Vector of variances}
}
\description{
Convert correlation matrix to covariance matrix
}
