% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvrt.r
\name{make_cor_mat}
\alias{make_cor_mat}
\title{Populate a correlation matrix}
\usage{
make_cor_mat(r, dim = 2)
}
\arguments{
\item{r}{Pearson's correlation to populate off-diagonals}

\item{dim}{Matrix dimensions}
}
\description{
Populate a correlation matrix
}
