% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvrt}
\alias{mvrt}
\title{Create a multivariate t-distributed random matrix}
\usage{
mvrt(n, mu, S, df = 1L)
}
\arguments{
\item{n}{Number of observations}

\item{mu}{Vector of sample means}

\item{S}{Covariance matrix}

\item{df}{Degrees of freedom}
}
\description{
Create a multivariate t-distributed random matrix
}
