% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvrt.r
\name{mvrtR}
\alias{mvrtR}
\title{Create a multivariate t-distributed random matrix}
\usage{
mvrtR(n, mu, S, df = n - 1, max_norm = 2, max_iterations = 1000,
  type_norm = "m")
}
\arguments{
\item{n}{Number of observations}

\item{mu}{Vector of sample means}

\item{S}{Covariance matrix}

\item{df}{Degrees of freedom}

\item{max_norm}{Maximum matrix norm}

\item{max_iterations}{Maximum number of iterations for the loop to search
for a matrix meeting \code{max_norm} criterion.}

\item{type_norm}{Type of matrix norm passed to \code{\link[base]{norm}}}
}
\description{
Create a multivariate t-distributed random matrix
}
