% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/mvrt.r
\name{mvrt}
\alias{mvrt}
\alias{mvrt2}
\alias{mvrt}
\title{Create a multivariate t-distributed random matrix}
\usage{
mvrt(n, mu, S, df = 1L)

mvrt2(n, mu, S, df = 1L, max_norm = 2, max_iterations = 1000L)
}
\arguments{
\item{n}{Number of observations.}

\item{mu}{Vector of sample means.}

\item{S}{Covariance matrix.}

\item{df}{Degrees of freedom.}

\item{max_norm}{Maximum matrix norm in difference between target correlation
matrix and the correlation matrix resulting from randomly generated data.
The greatest possible norm is 2. Use care with extemely small values, as
several iterations may be required to obtain such a matrix.}

\item{max_iterations}{Maximum number of iterations for the loop to search
for a matrix meeting \code{max_norm} criterion.}
}
\description{
These functions will create a multivariate t-distributed random matrix of a
specified mean and covariance structure.
}
\details{
The \code{mvrt2} fuction differs from \code{mvrt} in that it will ensure that the
resulting random data has a correlation matrix within a specified criterion.
No element of the resulting correlation matrix will deviate from the
corresponding element of the correlation matrix generated from
specifications by more than the absolute value of \code{max_norm}. This
checking comes at the expense of computational speed.
}
\examples{
S <- convert_R2S(make_cor_mat(.9),2:3)
x <- mvrt( 30, 4:5, S, 29)
y <- mvrt2(30, 4:5, S, 29, .01)

cor(x)
cor(y)
}
