% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvrt.r
\name{mvrtR}
\alias{mvrtR}
\title{Create a multivariate t-distributed random matrix}
\usage{
mvrtR(n, mu, S, df = n - 1, max.norm = Inf, type.norm = "m")
}
\arguments{
\item{n}{Number of observations}

\item{mu}{Vector of sample means}

\item{S}{Covariance matrix}

\item{df}{Degrees of freedom}

\item{max.norm}{Maximum matrix norm}

\item{type.norm}{Type of matrix norm passed to \code{\link[base]{norm}}}
}
\description{
Create a multivariate t-distributed random matrix
}
